Extract gitmodules discovery partial
